+++ src/main/resources/templates/index.html
@@ -107,7 +107,10 @@
       <p class="team__user-name">YOU</p>
     </div>
   </section>
-  <script src="/script/index.js"></script>
+  <script src="/script/util.js" type="module"></script>
+  <script src="/script/request.js" type="module"></script>
+  <script src="/script/command-button.js" type="module"></script>
+  <script src="/script/index.js" type="module"></script>
 </main>
 </body>
 </html>
